FIX :: feat#9에서 생긴 에러 긴급 수정
