+++ src/App.css
@@ -1,5 +1,13 @@
 .App {
   width: 100vi;
-  height: 100vh;
+  height: 100dvh;
   background-color: #1F2329;
+  display: flex;
+  flex-direction: row;
 }
+
+.App section {
+  padding: 1rem;
+  width: 76.75%;
+}
+
