+++ src/components/betting/betting-teams/RightTeam.jsx
@@ -23,7 +23,6 @@ function RightTeam(props){
 
     while (totalBettingMoney >= 1000){
         totalBettingMoney = totalBettingMoney / 1000;
-        console.log(totalBettingMoney);
         UnitIndex++;
     }
