+++ src/components/betting/betting-teams/bettingDetail.css
